refactor: separate hooks from useEventForm
