+++ package.json
@@ -56,5 +56,6 @@
     "vite": "^7.0.2",
     "vite-plugin-eslint": "^1.8.1",
     "vitest": "^3.2.4"
-  }
+  },
+  "packageManager": "pnpm@10.13.1+sha512.37ebf1a5c7a30d5fabe0c5df44ee8da4c965ca0c5af3dbab28c3a1681b70a256218d05c81c9c0dcf767ef6b8551eb5b960042b9ed4300c59242336377e01cfad"
 }
