+++ src/hooks/useEventEditor.ts
@@ -1,0 +1,11 @@
+import { Event } from '../types';
+import { useState } from 'react';
+
+export const useEventEditor = () => {
+  const [editingEvent, setEditingEvent] = useState<Event | null>(null);
+
+  return {
+    editingEvent,
+    setEditingEvent,
+  };
+};
