+++ src/hooks/useEventForm.ts
@@ -1,72 +1,63 @@
-import { ChangeEvent, useState } from 'react';
-
-import { Event, RepeatType } from '../types';
-import { getTimeErrorMessage } from '../utils/timeValidation';
-
-type TimeErrorRecord = Record<'startTimeError' | 'endTimeError', string | null>;
+import { Event } from '../types';
+import { useEventFormData } from './useEventFormData';
+import { useEventRepeatSettings } from './useEventRepeatSettings';
+import { useEventTimeManagement } from './useEventTimeManagement';
+import { useEventEditor } from './useEventEditor';
 
 export const useEventForm = (initialEvent?: Event) => {
-  const [title, setTitle] = useState(initialEvent?.title || '');
-  const [date, setDate] = useState(initialEvent?.date || '');
-  const [startTime, setStartTime] = useState(initialEvent?.startTime || '');
-  const [endTime, setEndTime] = useState(initialEvent?.endTime || '');
-  const [description, setDescription] = useState(initialEvent?.description || '');
-  const [location, setLocation] = useState(initialEvent?.location || '');
-  const [category, setCategory] = useState(initialEvent?.category || '업무');
-  const [isRepeating, setIsRepeating] = useState(initialEvent?.repeat.type !== 'none');
-  const [repeatType, setRepeatType] = useState<RepeatType>(initialEvent?.repeat.type || 'none');
-  const [repeatInterval, setRepeatInterval] = useState(initialEvent?.repeat.interval || 1);
-  const [repeatEndDate, setRepeatEndDate] = useState(initialEvent?.repeat.endDate || '');
-  const [notificationTime, setNotificationTime] = useState(initialEvent?.notificationTime || 10);
-
-  const [editingEvent, setEditingEvent] = useState<Event | null>(null);
-
-  const [{ startTimeError, endTimeError }, setTimeError] = useState<TimeErrorRecord>({
-    startTimeError: null,
-    endTimeError: null,
-  });
-
-  const handleStartTimeChange = (e: ChangeEvent<HTMLInputElement>) => {
-    const newStartTime = e.target.value;
-    setStartTime(newStartTime);
-    setTimeError(getTimeErrorMessage(newStartTime, endTime));
-  };
-
-  const handleEndTimeChange = (e: ChangeEvent<HTMLInputElement>) => {
-    const newEndTime = e.target.value;
-    setEndTime(newEndTime);
-    setTimeError(getTimeErrorMessage(startTime, newEndTime));
-  };
+  const {
+    title,
+    date,
+    description,
+    location,
+    category,
+    notificationTime,
+    resetEventForm,
+    updateEditEventForm,
+    setTitle,
+    setCategory,
+    setDate,
+    setLocation,
+    setNotificationTime,
+    setDescription,
+  } = useEventFormData(initialEvent);
+  const {
+    isRepeating,
+    repeatEndDate,
+    repeatInterval,
+    repeatType,
+    setIsRepeating,
+    setRepeatEndDate,
+    setRepeatInterval,
+    setRepeatType,
+    resetRepeat,
+    editRepeat,
+  } = useEventRepeatSettings(initialEvent);
+  const {
+    startTime,
+    endTime,
+    startTimeError,
+    endTimeError,
+    setEndTime,
+    setStartTime,
+    handleEndTimeChange,
+    handleStartTimeChange,
+    resetTime,
+    editTime,
+  } = useEventTimeManagement(initialEvent);
+  const { editingEvent, setEditingEvent } = useEventEditor();
 
   const resetForm = () => {
-    setTitle('');
-    setDate('');
-    setStartTime('');
-    setEndTime('');
-    setDescription('');
-    setLocation('');
-    setCategory('업무');
-    setIsRepeating(false);
-    setRepeatType('none');
-    setRepeatInterval(1);
-    setRepeatEndDate('');
-    setNotificationTime(10);
+    resetEventForm();
+    resetRepeat();
+    resetTime();
   };
 
   const editEvent = (event: Event) => {
     setEditingEvent(event);
-    setTitle(event.title);
-    setDate(event.date);
-    setStartTime(event.startTime);
-    setEndTime(event.endTime);
-    setDescription(event.description);
-    setLocation(event.location);
-    setCategory(event.category);
-    setIsRepeating(event.repeat.type !== 'none');
-    setRepeatType(event.repeat.type);
-    setRepeatInterval(event.repeat.interval);
-    setRepeatEndDate(event.repeat.endDate || '');
-    setNotificationTime(event.notificationTime);
+    updateEditEventForm(event);
+    editRepeat(event);
+    editTime(event);
   };
 
   return {
