+++ src/hooks/useEventRepeatSettings.ts
@@ -1,0 +1,36 @@
+import { useState } from 'react';
+import { Event, RepeatType } from '../types';
+
+export const useEventRepeatSettings = (initialEvent?: Event) => {
+  const [isRepeating, setIsRepeating] = useState(initialEvent?.repeat.type !== 'none');
+  const [repeatType, setRepeatType] = useState<RepeatType>(initialEvent?.repeat.type || 'none');
+  const [repeatInterval, setRepeatInterval] = useState(initialEvent?.repeat.interval || 1);
+  const [repeatEndDate, setRepeatEndDate] = useState(initialEvent?.repeat.endDate || '');
+
+  const resetRepeat = () => {
+    setIsRepeating(false);
+    setRepeatType('none');
+    setRepeatInterval(1);
+    setRepeatEndDate('');
+  };
+
+  const editRepeat = (event: Event) => {
+    setIsRepeating(event.repeat.type !== 'none');
+    setRepeatType(event.repeat.type);
+    setRepeatInterval(event.repeat.interval);
+    setRepeatEndDate(event.repeat.endDate || '');
+  };
+
+  return {
+    isRepeating,
+    setIsRepeating,
+    repeatType,
+    setRepeatType,
+    repeatInterval,
+    setRepeatInterval,
+    repeatEndDate,
+    setRepeatEndDate,
+    resetRepeat,
+    editRepeat,
+  };
+};
