+++ src/utils/date/calculations.ts
@@ -1,0 +1,50 @@
+/**
+ * 주어진 년도와 월의 일수를 반환합니다.
+ */
+export function getDaysInMonth(year: number, month: number): number {
+  return new Date(year, month, 0).getDate();
+}
+
+/**
+ * 주어진 날짜가 속한 주의 모든 날짜를 반환합니다.
+ */
+export function getWeekDates(date: Date): Date[] {
+  const day = date.getDay();
+  const diff = date.getDate() - day;
+  const sunday = new Date(date.setDate(diff));
+  const weekDates = [];
+  for (let i = 0; i < 7; i++) {
+    const nextDate = new Date(sunday);
+    nextDate.setDate(sunday.getDate() + i);
+    weekDates.push(nextDate);
+  }
+  return weekDates;
+}
+
+export function getWeeksAtMonth(currentDate: Date) {
+  const year = currentDate.getFullYear();
+  const month = currentDate.getMonth();
+  const daysInMonth = getDaysInMonth(year, month + 1);
+  const firstDayOfMonth = new Date(year, month, 1).getDay();
+  const days = Array.from({ length: daysInMonth }, (_, i) => i + 1);
+  const weeks = [];
+
+  const initWeek = () => Array(7).fill(null);
+
+  let week: Array<number | null> = initWeek();
+
+  for (let i = 0; i < firstDayOfMonth; i++) {
+    week[i] = null;
+  }
+
+  for (const day of days) {
+    const dayIndex = (firstDayOfMonth + day - 1) % 7;
+    week[dayIndex] = day;
+    if (dayIndex === 6 || day === daysInMonth) {
+      weeks.push(week);
+      week = initWeek();
+    }
+  }
+
+  return weeks;
+}
