+++ src/utils/date/formatting.ts
@@ -1,0 +1,40 @@
+export function formatWeek(targetDate: Date) {
+  const dayOfWeek = targetDate.getDay();
+  const diffToThursday = 4 - dayOfWeek;
+  const thursday = new Date(targetDate);
+  thursday.setDate(targetDate.getDate() + diffToThursday);
+
+  const year = thursday.getFullYear();
+  const month = thursday.getMonth() + 1;
+
+  const firstDayOfMonth = new Date(thursday.getFullYear(), thursday.getMonth(), 1);
+
+  const firstThursday = new Date(firstDayOfMonth);
+  firstThursday.setDate(1 + ((4 - firstDayOfMonth.getDay() + 7) % 7));
+
+  const weekNumber: number =
+    Math.floor((thursday.getTime() - firstThursday.getTime()) / (7 * 24 * 60 * 60 * 1000)) + 1;
+
+  return `${year}년 ${month}월 ${weekNumber}주`;
+}
+
+/**
+ * 주어진 날짜의 월 정보를 "YYYY년 M월" 형식으로 반환합니다.
+ */
+export function formatMonth(date: Date): string {
+  const year = date.getFullYear();
+  const month = date.getMonth() + 1;
+  return `${year}년 ${month}월`;
+}
+
+export function fillZero(value: number, size = 2) {
+  return String(value).padStart(size, '0');
+}
+
+export function formatDate(currentDate: Date, day?: number) {
+  return [
+    currentDate.getFullYear(),
+    fillZero(currentDate.getMonth() + 1),
+    fillZero(day ?? currentDate.getDate()),
+  ].join('-');
+}
