+++ src/utils/eventUtils.ts
@@ -1,5 +1,6 @@
 import { Event } from '../types';
-import { getWeekDates, isDateInRange } from './dateUtils';
+import { getWeekDates } from './date/calculations';
+import { isDateInRange } from './date/events';
 
 function filterEventsByDateRange(events: Event[], start: Date, end: Date): Event[] {
   return events.filter((event) => {
